Move tests around and add a test for lists for lists
